# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KnowFlow AI",
    "description": "Enterprise Knowledge & SOP Agent backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Hybrid Search",
        "description": "Executes hybrid (dense vector + sparse BM25) search across authorized company documents.\nServer-side authorization is strictly enforced based on user_context.",
        "operationId": "hybrid_search_api_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "description": "Search query string",
              "title": "Query"
            },
            "description": "Search query string"
          },
          {
            "name": "department_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional department filter",
              "title": "Department Id"
            },
            "description": "Optional department filter"
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Max number of results to return",
              "default": 10,
              "title": "Top K"
            },
            "description": "Max number of results to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evaluation/run": {
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Run Evaluation Benchmark",
        "description": "Triggers an automated evaluation benchmark across the golden test suite.\nCalculates Hit Rate @ 1/3/5, MRR, NDCG@5, Faithfulness, and reports\nFalse Positives and False Negatives separately.",
        "operationId": "run_evaluation_benchmark_api_evaluation_run_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Hybrid Search",
        "description": "Executes hybrid (dense vector + sparse BM25) search across authorized company documents.\nServer-side authorization is strictly enforced based on user_context.",
        "operationId": "hybrid_search_api_v1_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 500,
              "description": "Search query string",
              "title": "Query"
            },
            "description": "Search query string"
          },
          {
            "name": "department_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional department filter",
              "title": "Department Id"
            },
            "description": "Optional department filter"
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Max number of results to return",
              "default": 10,
              "title": "Top K"
            },
            "description": "Max number of results to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/evaluation/run": {
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Run Evaluation Benchmark",
        "description": "Triggers an automated evaluation benchmark across the golden test suite.\nCalculates Hit Rate @ 1/3/5, MRR, NDCG@5, Faithfulness, and reports\nFalse Positives and False Negatives separately.",
        "operationId": "run_evaluation_benchmark_api_v1_evaluation_run_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseEvaluationResult": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "is_out_of_scope": {
            "type": "boolean",
            "title": "Is Out Of Scope"
          },
          "retrieved_doc_titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Retrieved Doc Titles"
          },
          "retrieved_sections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Retrieved Sections"
          },
          "hit_at_1": {
            "type": "number",
            "title": "Hit At 1"
          },
          "hit_at_3": {
            "type": "number",
            "title": "Hit At 3"
          },
          "hit_at_5": {
            "type": "number",
            "title": "Hit At 5"
          },
          "mrr": {
            "type": "number",
            "title": "Mrr"
          },
          "ndcg_at_5": {
            "type": "number",
            "title": "Ndcg At 5"
          },
          "is_grounded": {
            "type": "boolean",
            "title": "Is Grounded"
          },
          "grounding_status": {
            "type": "string",
            "title": "Grounding Status"
          },
          "faithfulness_score": {
            "type": "number",
            "title": "Faithfulness Score"
          },
          "is_faithful": {
            "type": "boolean",
            "title": "Is Faithful"
          },
          "is_false_positive": {
            "type": "boolean",
            "title": "Is False Positive"
          },
          "is_false_negative": {
            "type": "boolean",
            "title": "Is False Negative"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "question",
          "category",
          "is_out_of_scope",
          "retrieved_doc_titles",
          "retrieved_sections",
          "hit_at_1",
          "hit_at_3",
          "hit_at_5",
          "mrr",
          "ndcg_at_5",
          "is_grounded",
          "grounding_status",
          "faithfulness_score",
          "is_faithful",
          "is_false_positive",
          "is_false_negative",
          "latency_ms",
          "passed"
        ],
        "title": "CaseEvaluationResult"
      },
...
      "EvaluationSummary": {
        "properties": {
          "total_cases": {
            "type": "integer",
            "title": "Total Cases"
          },
          "in_scope_cases": {
            "type": "integer",
            "title": "In Scope Cases"
          },
          "out_of_scope_cases": {
            "type": "integer",
            "title": "Out Of Scope Cases"
          },
          "passed_cases": {
            "type": "integer",
            "title": "Passed Cases"
          },
          "failed_cases": {
            "type": "integer",
            "title": "Failed Cases"
          },
          "pass_rate_pct": {
            "type": "number",
            "title": "Pass Rate Pct"
          },
          "hit_rate_at_1": {
            "type": "number",
            "title": "Hit Rate At 1"
          },
          "hit_rate_at_3": {
            "type": "number",
            "title": "Hit Rate At 3"
          },
          "hit_rate_at_5": {
            "type": "number",
            "title": "Hit Rate At 5"
          },
          "mrr": {
            "type": "number",
            "title": "Mrr"
          },
          "ndcg_at_5": {
            "type": "number",
            "title": "Ndcg At 5"
          },
          "avg_faithfulness": {
            "type": "number",
            "title": "Avg Faithfulness"
          },
          "false_positives": {
            "type": "integer",
            "title": "False Positives"
          },
          "false_negatives": {
            "type": "integer",
            "title": "False Negatives"
          },
          "refusal_accuracy_pct": {
            "type": "number",
            "title": "Refusal Accuracy Pct"
          },
          "avg_latency_ms": {
            "type": "number",
            "title": "Avg Latency Ms"
          },
          "case_results": {
            "items": {
              "$ref": "#/components/schemas/CaseEvaluationResult"
            },
            "type": "array",
            "title": "Case Results"
          }
        },
        "type": "object",
        "required": [
          "total_cases",
          "in_scope_cases",
          "out_of_scope_cases",
          "passed_cases",
          "failed_cases",
          "pass_rate_pct",
          "hit_rate_at_1",
          "hit_rate_at_3",
          "hit_rate_at_5",
          "mrr",
          "ndcg_at_5",
          "avg_faithfulness",
          "false_positives",
          "false_negatives",
          "refusal_accuracy_pct",
          "avg_latency_ms",
          "case_results"
        ],
        "title": "EvaluationSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchChunkResult": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "default": "text"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number",
            "default": 1
          },
          "section_heading": {
            "type": "string",
            "title": "Section Heading",
            "default": ""
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "dense_score": {
            "type": "number",
            "title": "Dense Score",
            "default": 0.0
          },
          "sparse_score": {
            "type": "number",
            "title": "Sparse Score",
            "default": 0.0
          },
          "rrf_score": {
            "type": "number",
            "title": "Rrf Score",
            "default": 0.0
          },
          "relevance_pct": {
            "type": "integer",
            "title": "Relevance Pct",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "document_title",
          "snippet"
        ],
        "title": "SearchChunkResult"
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "total_results": {
            "type": "integer",
            "title": "Total Results"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchChunkResult"
            },
            "type": "array",
            "title": "Results"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "query",
          "total_results"
        ],
        "title": "SearchResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}